requirements.txt
PyQt5
cryptography
psutil
pycrypto
